+++ lang/en.go
@@ -1,12 +1,19 @@
 package lang
 
 var EN_STRINGS = LangString{
-	"usage":            "CLI toot to use ai model",
-	"output-desc":      "Output directory",
-	"output-dir-empty": "Output directory is empty",
-	"silent":           "Disable printing log to stdout",
-	"no-args":          "No arguments provided",
-	"no-command":       "No command provided",
-	"text-usage":       "Generate text from a prompt",
-	"unknown-sdk":      "Unknown sdk \"%s\"",
+	"usage":              "CLI toot to use ai model",
+	"output-desc":        "Output directory",
+	"output-dir-empty":   "Output directory is empty",
+	"silent":             "Disable printing log to stdout",
+	"no-args":            "No arguments provided",
+	"no-command":         "No command provided",
+	"unknown-sdk":        "Unknown sdk \"%s\"",
+	"sdk-model-usage":    "Select a model",
+	"text-usage":         "Generate text from a prompt",
+	"text-temp-usage":    "Set temperature",
+	"text-system-usage":  "Instruction with role system (use \"-\" for stdin)",
+	"text-history-usage": "Select a history",
+	"text-clear-usage":   "Clear history",
+	"type-required":      "Type is required",
+	"apiKey-required":    "API key is required",
 }