+++ lang/fr.go
@@ -1,12 +1,19 @@
 package lang
 
 var FR_STRINGS = LangString{
-	"usage":            "CLI pour utiliser des modèles d'IA",
-	"output-desc":      "Répertoire de sortie",
-	"output-dir-empty": "Le répertoire de sortie est vide",
-	"silent":           "Désactiver l'impression du journal sur stdout",
-	"no-args":          "Aucun argument fourni",
-	"no-command":       "Aucune commande fournie",
-	"text-usage":       "Générer du texte à partir d'un prompt",
-	"unknown-sdk":      "Sdk inconnu \"%s\"",
+	"usage":              "CLI pour utiliser des modèles d'IA",
+	"output-desc":        "Répertoire de sortie",
+	"output-dir-empty":   "Le répertoire de sortie est vide",
+	"silent":             "Désactiver l'impression du journal sur stdout",
+	"no-args":            "Aucun argument fourni",
+	"no-command":         "Aucune commande fournie",
+	"unknown-sdk":        "Sdk inconnu \"%s\"",
+	"sdk-model-usage":    "Sélectionner un modèle",
+	"text-usage":         "Générer du texte à partir d'un prompt",
+	"text-temp-usage":    "Définir la température",
+	"text-system-usage":  "Instruction avec rôle système (utilisez \"-\" pour stdin)",
+	"text-history-usage": "Sélectionner un historique",
+	"text-clear-usage":   "Effacer l'historique",
+	"type-required":      "Le type est requis",
+	"apiKey-required":    "La clé API est requise",
 }