+++ service/text.go
@@ -1,16 +1,20 @@
 package service
 
 import (
+	"fmt"
+
 	"github.com/LordPax/aicli/sdk"
 )
 
 func SendTextRequest(prompt string) error {
 	textSdk := sdk.GetSdkText()
 
-	_, err := textSdk.SendRequest(prompt)
+	resp, err := textSdk.SendRequest(prompt)
 	if err != nil {
 		return err
 	}
+
+	fmt.Println(resp.Content)
 
 	return nil
 }