+++ utils/utils.go
@@ -1,6 +1,8 @@
 package utils
 
 import (
+	"bytes"
+	"net/http"
 	"os"
 )
 
@@ -11,4 +13,22 @@ func FileExist(file string) bool {
 
 func RmTmpDir() error {
 	return os.RemoveAll(os.Getenv("TMP_DIR"))
+}
+
+func PostRequest(url string, data []byte, option map[string]string) (*http.Response, error) {
+	req, err := http.NewRequest("POST", url, bytes.NewBuffer(data))
+	if err != nil {
+		return nil, err
+	}
+
+	for key, value := range option {
+		req.Header.Set(key, value)
+	}
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	return resp, nil
 }