+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
-* Add sdk for openai text completion
+* Add parameter `history`, `model`, `temp`, `system` for text command
+* Add history with name for text command
+* Add sdk openai for text command
 * Add localized for `en` and `fr`
 * Read config.ini file